chore(web): refactor the workflow settings/snippet sidebar
