+++ apps/web/src/pages/templates/components/ProvidersPage.tsx
@@ -7,10 +7,10 @@ import { IntegrationsStoreModal } from '../../integrations/IntegrationsStoreModa
 import type { IIntegratedProvider } from '../../integrations/types';
 import { useProviders } from '../../integrations/useProviders';
 import { ListProviders } from './ListProviders';
-import { SubPageWrapper } from './SubPageWrapper';
 import { WorkflowSettingsTabs } from './WorkflowSettingsTabs';
 import { useIsMultiProviderConfigurationEnabled } from '../../../hooks';
 import { IntegrationsListModal } from '../../integrations/IntegrationsListModal';
+import { WorkflowSidebar } from './WorkflowSidebar';
 
 export function ProvidersPage() {
   const { emailProviders, smsProvider, chatProvider, pushProvider, inAppProvider, isLoading } = useProviders();
@@ -33,13 +33,7 @@ export function ProvidersPage() {
 
   return (
     <>
-      <SubPageWrapper
-        title="Workflow Settings"
-        style={{
-          display: 'flex',
-          flexFlow: 'column',
-        }}
-      >
+      <WorkflowSidebar title="Workflow Settings">
         <WorkflowSettingsTabs />
         <ScrollArea h="calc(100vh - 220px)" offsetScrollbars mr={-12}>
           <ListProviders
@@ -73,7 +67,7 @@ export function ProvidersPage() {
             providers={smsProvider}
           />
         </ScrollArea>
-      </SubPageWrapper>
+      </WorkflowSidebar>
       {isMultiProviderConfigurationEnabled ? (
         <IntegrationsListModal
           isOpen={configureChannel !== undefined}
