+++ apps/web/src/pages/templates/components/SnippetPage.tsx
@@ -1,16 +1,17 @@
 import { Text } from '@mantine/core';
+
 import { colors } from '../../../design-system';
 import { TriggerSnippetTabs } from './TriggerSnippetTabs';
 import { useTemplateEditorForm } from './TemplateEditorFormProvider';
-import { SubPageWrapper } from './SubPageWrapper';
 import { TriggerSegmentControl } from './TriggerSegmentControl';
 import { When } from '../../../components/utils/When';
+import { WorkflowSidebar } from './WorkflowSidebar';
 
 export function SnippetPage() {
   const { trigger, isCreating, isUpdating } = useTemplateEditorForm();
 
   return (
-    <SubPageWrapper title="Trigger">
+    <WorkflowSidebar title="Trigger">
       <Text color={colors.B60} mt={-16}>
         Test trigger as if you sent it from your API or implement it by copy/pasting it into the codebase of your
         application.
@@ -19,6 +20,6 @@ export function SnippetPage() {
         <TriggerSegmentControl />
         {trigger && <TriggerSnippetTabs trigger={trigger} />}
       </When>
-    </SubPageWrapper>
+    </WorkflowSidebar>
   );
 }
