+++ apps/web/src/pages/templates/components/TestWorkflow.tsx
@@ -4,8 +4,9 @@ import { useForm } from '@mantine/form';
 import { useMutation } from '@tanstack/react-query';
 import * as Sentry from '@sentry/react';
 import * as capitalize from 'lodash.capitalize';
-
+import { useDisclosure } from '@mantine/hooks';
 import { IUserEntity, INotificationTriggerVariable } from '@novu/shared';
+
 import { Button, colors } from '../../../design-system';
 import { inputStyles } from '../../../design-system/config/inputs.styles';
 import { errorMessage, successMessage } from '../../../utils/notifications';
@@ -13,9 +14,8 @@ import { useAuthContext } from '../../../components/providers/AuthProvider';
 import { getSubscriberValue, getPayloadValue } from './TriggerSnippetTabs';
 import { testTrigger } from '../../../api/notification-templates';
 import { ExecutionDetailsModalWrapper } from './ExecutionDetailsModalWrapper';
-import { useDisclosure } from '@mantine/hooks';
-import { SubPageWrapper } from './SubPageWrapper';
 import { TriggerSegmentControl } from './TriggerSegmentControl';
+import { WorkflowSidebar } from './WorkflowSidebar';
 
 const makeToValue = (subscriberVariables: INotificationTriggerVariable[], currentUser?: IUserEntity) => {
   const subsVars = getSubscriberValue(
@@ -115,13 +115,12 @@ export function TestWorkflow({ trigger }) {
 
   return (
     <>
-      <SubPageWrapper title="Trigger">
+      <WorkflowSidebar title="Trigger">
         <Text color={colors.B60} mt={-16}>
           Test trigger as if you sent it from your API or implement it by copy/pasting it into the codebase of your
           application.
         </Text>
         <TriggerSegmentControl />
-
         <JsonInput
           data-test-id="test-trigger-to-param"
           formatOnBlur
@@ -165,7 +164,7 @@ export function TestWorkflow({ trigger }) {
             validationError="Invalid JSON"
           />
         ))}
-        <Group position="right" mt={'auto'}>
+        <Group position="right" mt={'auto'} mb={24}>
           <div data-test-id="test-workflow-btn">
             <Button
               sx={{
@@ -184,7 +183,7 @@ export function TestWorkflow({ trigger }) {
             </Button>
           </div>
         </Group>
-      </SubPageWrapper>
+      </WorkflowSidebar>
       <ExecutionDetailsModalWrapper
         transactionId={transactionId}
         isOpen={executionModalOpened}
