+++ apps/web/src/pages/templates/components/WorkflowSidebar.tsx
@@ -1,0 +1,38 @@
+import { ReactNode } from 'react';
+import { Title } from '@mantine/core';
+import { useNavigate } from 'react-router-dom';
+
+import { colors, Sidebar } from '../../../design-system';
+import { useBasePath } from '../hooks/useBasePath';
+
+export const WorkflowSidebar = ({ children, title }: { children: ReactNode; title: string }) => {
+  const navigate = useNavigate();
+  const path = useBasePath();
+
+  return (
+    <Sidebar
+      isOpened
+      isParentScrollable
+      customHeader={
+        <Title
+          sx={{
+            lineHeight: '48px',
+            flex: '1 1 auto',
+            display: 'flex',
+            flexFlow: 'Column',
+          }}
+          color={colors.white}
+          size={20}
+          weight="bold"
+        >
+          {title}
+        </Title>
+      }
+      onClose={() => {
+        navigate(path);
+      }}
+    >
+      {children}
+    </Sidebar>
+  );
+};
